# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Palplus UAA",
        "description": "Palplus Swagger API description",
        "contact": {
            "email": "michaelkinare@gmail.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/",
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "All Users List",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Retrieve User",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Soft delete a user",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Update a user",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "country_code",
                        "in": "path",
                        "description": "name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "password",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/{id}/force": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Force delete a user",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/activate/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Activate a user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/deactivate/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Dectivate a user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/wallet": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "My wallet",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "My Profile",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/nok": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "My Next of Kin",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/groups": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "My Groups",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/transactions": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "My Transactions",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/accounts": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "My Accounts",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/payments": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "My Payments",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/loan": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "My Loan",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
...
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "description": "Auth Bearer Token Format as 'Bearer <access_token>'",
            "name": "Authorization",
            "in": "header"
        }
    }
}